+++ src/LuauExt.cpp
@@ -481,6 +481,8 @@ std::string toStringNamedFunction(Luau::ModulePtr module, const Luau::FunctionTy
     opts.functionTypeArguments = true;
     opts.hideNamedFunctionTypeParameters = false;
     opts.hideTableKind = stringOpts.hideTableKind;
+    opts.useLineBreaks = stringOpts.multiline;
+    opts.indent = stringOpts.multiline;
     if (scope)
         opts.scope = *scope;
     auto functionString = Luau::toStringNamedFunction("", *ftv, opts);